Rename CheckBeforeSending.jsx to CheckBeforeSendingStep.jsx
